Don't attempt to destructure sets
Precision doesn't actually tell us about the type of a CHAR column. It
simply tells us how many bytes the column has to hold data, which will
vary for CHAR columns as well as ENUM and SET columns.

We simply need to respond with the string and allow the caller to split
results themselves.